Split reducer into chunks
