+++ project/src/store/reducer.ts
@@ -1,25 +1,8 @@
-import { createReducer } from '@reduxjs/toolkit';
-import { selectCity, loadOffers, setAuth } from './action';
-import { City } from '../const';
-import { Offer } from '../types/offer';
+import { combineReducers } from '@reduxjs/toolkit';
+import { data } from './data/data';
+import { user } from './user/user';
 
-const initialState = {
-  city: City.Paris,
-  offers: [] as Offer[],
-  loading: true,
-  authStatus: false,
-};
-
-export const reducer = createReducer(initialState, (builder) => {
-  builder
-    .addCase(selectCity, (state, { payload: { city } }) => {
-      state.city = city;
-    })
-    .addCase(loadOffers, (state, { payload: { offers } }) => {
-      state.offers = offers;
-      state.loading = false;
-    })
-    .addCase(setAuth, (state, { payload: { auth } }) => {
-      state.authStatus = auth;
-    });
+export const reducer = combineReducers({
+  data: data.reducer,
+  user: user.reducer,
 });
